move json array DBFunctions to functions namespace
