+++ db/tests/types/test_json.py
@@ -1,24 +1,27 @@
 import pytest
-from db.types.custom import json_array
 from db.utils import execute_pg_query
-from db.functions.base import ColumnName, Noop
+from db.functions.base import ColumnName, Noop, JsonArrayContains
+from db.functions.packed import (
+    JsonLengthEquals, JsonLengthGreaterThan, JsonLengthGreaterorEqual,
+    JsonLengthLessThan, JsonLengthLessorEqual, JsonNotEmpty
+)
 from db.functions.operations.apply import apply_db_function_as_filter
 
 
 @pytest.mark.parametrize("main_db_function,literal_param,expected_count", [
-    (json_array.JsonLengthEquals, 0, 1),
-    (json_array.JsonLengthEquals, 3, 11),
-    (json_array.JsonLengthEquals, 4, 1),
-    (json_array.JsonLengthEquals, 6, 1),
-    (json_array.JsonLengthGreaterThan, 3, 2),
-    (json_array.JsonLengthGreaterorEqual, 3, 13),
-    (json_array.JsonLengthLessThan, 4, 12),
-    (json_array.JsonLengthLessorEqual, 4, 13),
-    (json_array.JsonLengthGreaterorEqual, 4, 2),
-    (json_array.JsonArrayContains, [1, 2, 3], 1),
-    (json_array.JsonArrayContains, ["Ford", "BMW"], 6),
-    (json_array.JsonArrayContains, ["BMW", "Ford", "Fiat", "Fiat"], 3),
-    (json_array.JsonNotEmpty, '', 13),
+    (JsonLengthEquals, 0, 1),
+    (JsonLengthEquals, 3, 11),
+    (JsonLengthEquals, 4, 1),
+    (JsonLengthEquals, 6, 1),
+    (JsonLengthGreaterThan, 3, 2),
+    (JsonLengthGreaterorEqual, 3, 13),
+    (JsonLengthLessThan, 4, 12),
+    (JsonLengthLessorEqual, 4, 13),
+    (JsonLengthGreaterorEqual, 4, 2),
+    (JsonArrayContains, [1, 2, 3], 1),
+    (JsonArrayContains, ["Ford", "BMW"], 6),
+    (JsonArrayContains, ["BMW", "Ford", "Fiat", "Fiat"], 3),
+    (JsonNotEmpty, '', 13),
 ])
 def test_json_array_filter_functions(json_table_obj, main_db_function, literal_param, expected_count):
     table, engine = json_table_obj
